fix: CostOverrideResponse usa float em vez de int para evitar falha de validação Pydantic
fantasy_cost dos jogadores pode ser decimal (ex: 19.53). O schema declarava
Optional[int], fazendo o Pydantic v2 rejeitar a resposta com ValidationError (500)
ao tentar salvar um override. Também melhora exibição de erros de validação no
frontend (detail como array → string legível).

+++ app/routers/admin/rosters.py
@@ -42,9 +42,9 @@ class CostOverrideRequest(BaseModel):
 
 class CostOverrideResponse(BaseModel):
     roster_id: int
-    fantasy_cost: Optional[int]
-    cost_override: Optional[int]
-    effective_cost: Optional[int]
+    fantasy_cost: Optional[float]
+    cost_override: Optional[float]
+    effective_cost: Optional[float]
 
     model_config = {"from_attributes": True}
